Moved WebView and Xaml Frame into a single control
