+++ UserControls/WebXamlView.xaml
@@ -1,0 +1,15 @@
+﻿<UserControl
+    x:Class="Pulse_Browser.UserControls.WebXamlView"
+    xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+    xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+    xmlns:local="using:Pulse_Browser.UserControls"
+    xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
+    xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
+    xmlns:helpers="using:Pulse_Browser.Helpers"
+    mc:Ignorable="d">
+
+    <Grid>
+        <Frame x:Name="AppFrame" HorizontalAlignment="Stretch" VerticalAlignment="Stretch" />
+        <WebView x:Name="AppWebView" Visibility="{x:Bind helpers:Converters.BooleanToVisibility(ViewModel.WebViewShown), Mode=OneWay}" Source="{x:Bind ViewModel.CurrentAddress, Mode=TwoWay}" />
+    </Grid>
+</UserControl>
